Added checks and new commands
We ensure that all embeds stored in the csv file has something to display such that it doesn't break the discord. Since anywhere else we either already have an embed in the message to follow from, or we create our own embed (where we enforce a title)

Also created the skeleton of the add_new_ruleset command, where a description is optional but title is mandatory.